+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@
 * Fixes handling of relative imports for WDL specified using non-file URIs (e.g. http(s))
 * Fixes quoting and escaping of import URIs by code formatter/generator
 
-## 0.13.0
+## 0.13.0 (2021-05-20)
 
 * **Breaking changes**:
     * `loc` is moved from the first to second parameter section in all AST case classes to prevent it from being included in the `equals` and `hashCode` implementations